[Feature] 내 모집글 API 연동 및 로그인 상태 체크 추가

+++ feature/recruitment/src/main/java/in/koreatech/koin/feature/recruitment/ui/notification/RecruitmentNotificationScreen.kt
@@ -50,7 +50,7 @@ import org.orbitmvi.orbit.compose.collectSideEffect
 @Composable
 internal fun RecruitmentNotificationScreen(
     onBack: () -> Unit = {},
-    onNavigateToApplicantManagement: (Long) -> Unit = {},
+    onNavigateToApplicantManagement: (Int) -> Unit = {},
     viewModel: RecruitmentNotificationViewModel = hiltViewModel()
 ) {
     val uiState by viewModel.collectAsState()
